requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==5.1.3  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.4.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-webpack-loader==3.1.1  # https://github.com/django-webpack/django-webpack-loader


django-classy-tags==4.1.0
django-cms==4.1.4
django-filer==3.2.3
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-storage-url==0.12.1
django-storages==1.14.4
django-treebeard==4.7.1
# cms
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-frontend==1.3.5
djangocms-text==0.5.3
djangocms-versioning==2.1.0


easy-thumbnails[svg]==2.10
furl==2.1.3
jmespath==1.0.1
lxml==5.3.0
nh3==0.2.18
orderedmultidict==1.0.1
packaging==24.2
pip-tools==7.4.1
psycopg2==2.9.10
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
reportlab==4.2.5
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
typing-extensions==4.12.2
urllib3==2.2.3
uwsgi==2.0.28
webencodings==0.5.1
wheel==0.45.0

djangocms-seo==0.1.2
django-comments-xtd==2.10.2
djangocms-text-ckeditor5==0.2.0
django-ckeditor
djangocms-simple-admin-style==1.1.1